Ensemble, RunPlan(Vector) testing and RNG Changes (#665)
* Switch to uint64_t seeds and std::MT19937_64 for host RNG.

This makes the full range of entropy available for curand, and upgrades host rng engines to the 64 bit version of mt19937

* New Feature: Implement RunPlanVector::getRandomPropertySeed

Provides access to the (initally) randomly generated seed used to generate new properties on the host.

The returned value was  only valid for properties generated since the seed was last set, so use is limited

* DocFix: Remove invalid statement in EnsembleConfig::&Config()

* Bugfix: RunPlanVector::operator* return value

* Bugfix: RunPlanVector::setProperty element comparison fix

* Bugfix: Missing single quote in exception message

* Bugfix: CUDASimulation.processExitLog was using step_log_config not exit_log_config

* Bugfix: Do not wrap ensemble::getConfig or simulation::getCUDAConfig

These methods are supposed to return immutable configuraiton structs, but python/swig wrapping results in a mutable object.

* Bugfix: Invert comparisons when getting properties from RunPlans

This was using the result of unordered_map::find when it did not find the requested name, rather than the other way around.

* Bugfix: RunPlan(Vector) get/setProperty element index comparison

Also fixes a number of exception strings

* Bugfix: Change default RunPlan steps to 1
This matches default uses, and avoids the default value potentially throwing exceptions

* Bugfix: Exception type change(s) in RunPlan::setPropertyArray

These now match the std::array version

* Bugfix: Python RunPlanVector do not wrap setPropertyRandom

* Bugfix: Python RunPlanVector Do not rename setProperty*Random to not match the c++api

* Windows CI: If workflow_dispatch is used, build tests if VISUALISATION=OFF

* Disable FLAMEGPU Version output during CUDAEnsemble initialisation

This occured before it could be disabled programatically

* Tests: Add tests for uint64_t RNG seeds, RunPlan, RunPlanVector and CUDAEnsemble

Some python tests are skipped due to current implementation not allowing them

+++ include/flamegpu/gpu/CUDASimulation.h
@@ -298,7 +298,7 @@ class CUDASimulation : public Simulation {
      * Reinitalises random generation for this model and all submodels
      * @param seed New random seed (this updates stored seed in config)
      */
-    void reseed(const unsigned int &seed);
+    void reseed(const uint64_t &seed);
     /**
      * Number of times step() has been called since sim was last reset/init
      */
